fix(review): derive failure floor from canonical renderer
